wip: relocate mfm planner backend, extract shared utilities
