+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/mfm/__init__.py
@@ -1,0 +1,5 @@
+"""Legacy MFM planner backend. Deletable when MotionBricks fully replaces it."""
+
+from robotic_grounding.planner.mfm.inference import MotionInferenceAgent
+
+__all__ = ["MotionInferenceAgent"]
