+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/mfm/assets/skeleton/g1_skel_Tpose.bvh
