+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/mfm/assets/skeleton/skeleton.pkl
