+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/mfm/assets/skeleton/standard_t_pose_g1skel32_joint_quat.p
