+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/mfm/assets/skeleton/t_pose_offset_standard.npy
