+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/mfm/assets/skeletons/g1skel32/standard_t_pose_g1skel32_joint_quat.p
