+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/mfm/assets/stats/local_mean.npy
