+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/mfm/assets/stats/root_global_mean.npy
