+++ robotic_grounding/source/robotic_grounding/robotic_grounding/planner/mfm/assets/stats/root_local_mean.npy
